refactor: move external-ontology generators into importers/ subpackage
