+++ app/sentier_vocab/importers/__init__.py
@@ -1,0 +1,1 @@
+"""External-ontology importers (transitional layer)."""
